+++ e2e/integration_test.go
@@ -605,6 +605,7 @@ func testNetworkDelay(t *testing.T, m *e2e.Minikube, e *e2e.Extension) {
 		ip                  []string
 		hostname            []string
 		port                []string
+		excludeIp           []string
 		interfaces          []string
 		restrictedEndpoints []action_kit_api.RestrictedEndpoint
 		wantedDelay         bool
@@ -638,6 +639,12 @@ func testNetworkDelay(t *testing.T, m *e2e.Minikube, e *e2e.Extension) {
 			restrictedEndpoints: generateRestrictedEndpoints(1500),
 			wantedDelay:         true,
 		},
+		{
+			name:                "should not delay traffic for excluded cidr",
+			excludeIp:           []string{fmt.Sprintf("%s/32", netperf.ServerIp)},
+			restrictedEndpoints: generateRestrictedEndpoints(1500),
+			wantedDelay:         false,
+		},
 	}
 
 	unaffectedLatency, err := netperf.MeasureLatency()
@@ -652,6 +659,7 @@ func testNetworkDelay(t *testing.T, m *e2e.Minikube, e *e2e.Extension) {
 			"ip":                 tt.ip,
 			"hostname":           tt.hostname,
 			"port":               tt.port,
+			"excludeIp":          tt.excludeIp,
 			"networkInterface":   tt.interfaces,
 		}